+++ sdk/communication/Azure.Communication.CallAutomation/src/CallConnection.cs
@@ -274,9 +274,12 @@ namespace Azure.Communication.CallAutomation
 
         private static TransferToParticipantRequestInternal CreateTransferToParticipantRequest(TransferToParticipantOptions options)
         {
-            TransferToParticipantRequestInternal request = new TransferToParticipantRequestInternal(CommunicationIdentifierSerializer.Serialize(options.Target));
-
-            request.OperationContext = options.OperationContext == default ? Guid.NewGuid().ToString() : options.OperationContext;
+            TransferToParticipantRequestInternal request = new(CommunicationIdentifierSerializer.Serialize(options.Target))
+            {
+                OperationContext = options.OperationContext == default ? Guid.NewGuid().ToString() : options.OperationContext,
+                Transferee = options.Transferee == default ? null : CommunicationIdentifierSerializer.Serialize(options.Transferee),
+                OverrideCallbackUri = options.OverrideCallbackUrl
+            };
 
             return request;
         }
@@ -376,10 +379,9 @@ namespace Azure.Communication.CallAutomation
                     ? null
                     : new PhoneNumberIdentifierModel(options.ParticipantToAdd.SourceCallerIdNumber.PhoneNumber),
                 SourceDisplayName = options.ParticipantToAdd.SourceDisplayName,
-                OperationContext = options.OperationContext == default ? Guid.NewGuid().ToString() : options.OperationContext
+                OperationContext = options.OperationContext == default ? Guid.NewGuid().ToString() : options.OperationContext,
+                InvitationTimeoutInSeconds = options.InvitationTimeoutInSeconds
             };
-
-            request.InvitationTimeoutInSeconds = options.InvitationTimeoutInSeconds;
 
             return request;
         }
@@ -523,9 +525,11 @@ namespace Azure.Communication.CallAutomation
                 // validate RequestInitiator is not null or empty
                 Argument.AssertNotNull(options.ParticipantToRemove, nameof(options.ParticipantToRemove));
 
-                RemoveParticipantRequestInternal request = new(CommunicationIdentifierSerializer.Serialize(options.ParticipantToRemove));
-
-                request.OperationContext = options.OperationContext == default ? Guid.NewGuid().ToString() : options.OperationContext;
+                RemoveParticipantRequestInternal request = new(CommunicationIdentifierSerializer.Serialize(options.ParticipantToRemove))
+                {
+                    OperationContext = options.OperationContext == default ? Guid.NewGuid().ToString() : options.OperationContext,
+                    OverrideCallbackUri = options.OverrideCallbackUrl
+                };
 
                 var response = await RestClient.RemoveParticipantAsync(CallConnectionId, request, cancellationToken).ConfigureAwait(false);
 
@@ -571,9 +575,11 @@ namespace Azure.Communication.CallAutomation
                 if (options == null)
                     throw new ArgumentNullException(nameof(options));
 
-                RemoveParticipantRequestInternal request = new(CommunicationIdentifierSerializer.Serialize(options.ParticipantToRemove));
-
-                options.OperationContext = options.OperationContext == default ? Guid.NewGuid().ToString() : options.OperationContext;
+                RemoveParticipantRequestInternal request = new(CommunicationIdentifierSerializer.Serialize(options.ParticipantToRemove))
+                {
+                    OperationContext = options.OperationContext == default ? Guid.NewGuid().ToString() : options.OperationContext,
+                    OverrideCallbackUri = options.OverrideCallbackUrl,
+                };
 
                 var response = RestClient.RemoveParticipant(CallConnectionId, request, cancellationToken);
 
@@ -664,7 +670,6 @@ namespace Azure.Communication.CallAutomation
                     CallConnectionId,
                     request,
                     cancellationToken).ConfigureAwait(false);
-
                 return Response.FromValue(new MuteParticipantResult(response.Value), response.GetRawResponse());
             }
             catch (Exception ex)
@@ -674,13 +679,13 @@ namespace Azure.Communication.CallAutomation
             }
         }
 
-        /// <summary>
-        /// Cancel add participant operation.
-        /// </summary>
-        /// <param name="invitationId"></param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        public virtual Task<Response<CancelAddParticipantResult>> CancelAddParticipantAsync(string invitationId, CancellationToken cancellationToken = default)
+    /// <summary>
+    /// Cancel add participant operation.
+    /// </summary>
+    /// <param name="invitationId"></param>
+    /// <param name="cancellationToken"></param>
+    /// <returns></returns>
+    public virtual Task<Response<CancelAddParticipantResult>> CancelAddParticipantAsync(string invitationId, CancellationToken cancellationToken = default)
         {
             return CancelAddParticipantAsync(new CancelAddParticipantOptions(invitationId), cancellationToken);
         }
@@ -706,7 +711,7 @@ namespace Azure.Communication.CallAutomation
                 var request = new CancelAddParticipantRequestInternal(options.InvitationId)
                 {
                     OperationContext = options.OperationContext == default ? Guid.NewGuid().ToString() : options.OperationContext,
-                    CallbackUri = options.CallbackUri?.AbsoluteUri,
+                    OverrideCallbackUri = options.OverrideCallbackUrl
                 };
                 var response = await RestClient.CancelAddParticipantAsync(CallConnectionId, request, cancellationToken).ConfigureAwait(false);
                 var result = new CancelAddParticipantResult(response);
@@ -753,7 +758,7 @@ namespace Azure.Communication.CallAutomation
                 var request = new CancelAddParticipantRequestInternal(options.InvitationId)
                 {
                     OperationContext = options.OperationContext,
-                    CallbackUri = options.CallbackUri?.AbsoluteUri,
+                    OverrideCallbackUri = options.OverrideCallbackUrl
                 };
                 var response = RestClient.CancelAddParticipant(CallConnectionId, request, cancellationToken);
                 var result = new CancelAddParticipantResult(response);