+++ sdk/communication/Azure.Communication.CallAutomation/CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 1.1.0 (Unreleased)
 
 ### Features Added
-- StartRecording now accepts PauseOnStart.
+- Mid-Call actions support overriding callback uri 
+- Cancel adding Participant invitation 
+- Support transfer a participant in a group call to another participant
 
 ### Breaking Changes